Reorganize existing code
