+++ pkg/twitch/eventsubWebsocketClient.go
